requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.9.1.post1
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo;python_version<"3.9"
beautifulsoup4==4.11.1
bilibili-api-python==13.3.0
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
decorator==4.4.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
keyboard==0.13.5
lxml==4.9.1
moviepy==1.0.3
multidict==6.0.2
numpy==1.23.5
Pillow==9.2.0
proglog==0.1.10
pyasn1==0.4.8
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
sanic==22.9.1
sanic-routing==22.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
websockets==10.4
yarl==1.8.1

schedule~=1.2.1